--- src/movable.cpp
@@ -1,1 +1,0 @@
-#include "movable.hpp"
